nes: refactor edit-filtering and fix non-dynamic whitespace-only edit filtering
